scripts/train_model.py
#!/usr/bin/env python3
"""
Training script for AI Space Exploration models.

Usage:
    python scripts/train_model.py --model satellite_classifier --epochs 50
    python scripts/train_model.py --model object_detector --batch-size 32
    python scripts/train_model.py --model navigation_agent --episodes 1000
"""

import argparse
import torch
import torch.nn as nn
import torch.optim as optim
from pathlib import Path
import yaml
from typing import Dict
import sys

from src.models import (
    SatelliteImageClassifier,
    AstronomicalDetector,
    NavigationAgent,
    GalaxyClassifier,
    ExoplanetDetector
)


def parse_args():
    """Parse command line arguments."""
    parser = argparse.ArgumentParser(
        description="Train AI Space Exploration models"
    )
    
    parser.add_argument(
        "--model",
        type=str,
        required=True,
        choices=[
            'satellite_classifier',
            'object_detector',
            'navigation_agent',
            'galaxy_classifier',
            'exoplanet_detector'
        ],
        help="Model to train"
    )
    
    parser.add_argument(
        "--config",
        type=str,
        default=None,
        help="Path to config file"
    )
    
    parser.add_argument(
        "--epochs",
        type=int,
        default=50,
        help="Number of training epochs"
    )
    
    parser.add_argument(
        "--batch-size",
        type=int,
        default=32,
        help="Batch size for training"
    )
    
    parser.add_argument(
        "--lr",
        type=float,
        default=0.001,
        help="Learning rate"
    )
    
    parser.add_argument(
        "--output-dir",
        type=str,
        default="models",
        help="Output directory for checkpoints"
    )
    
    parser.add_argument(
        "--device",
        type=str,
        default="cuda" if torch.cuda.is_available() else "cpu",
        help="Device to train on"
    )
    
    parser.add_argument(
        "--resume",
        type=str,
        default=None,
        help="Path to checkpoint to resume from"
    )
    
    return parser.parse_args()


def load_config(config_path: str) -> Dict:
    """Load configuration from YAML file."""
    if config_path and Path(config_path).exists():
        with open(config_path, 'r') as f:
            return yaml.safe_load(f)
    return {}


def train_satellite_classifier(args, config):
    """Train satellite image classifier."""
    print("🛰️ Training Satellite Image Classifier...")
    
    # Initialize model
    model = SatelliteImageClassifier(
        num_classes=config.get('num_classes', 10)
    )
    model = model.to(args.device)
    
    # Training loop (simplified)
    print(f"✓ Model initialized with {sum(p.numel() for p in model.parameters()):,} parameters")
    print(f"✓ Training on {args.device}")
    print(f"✓ Epochs: {args.epochs}, Batch size: {args.batch_size}, LR: {args.lr}")
    
    # Placeholder training
    print("\n🏋️ Training in progress...")
    print("Note: This is a placeholder. Add actual training logic with your dataset.")
    
    # Save model
    output_path = Path(args.output_dir) / f"{args.model}_final.pth"
    output_path.parent.mkdir(exist_ok=True)
    torch.save({
        'model_state_dict': model.state_dict(),
        'config': config,
    }, output_path)
    
    print(f"\n✅ Training complete! Model saved to {output_path}")


def train_object_detector(args, config):
    """Train astronomical object detector."""
    print("🔭 Training Astronomical Object Detector...")
    
    model = AstronomicalDetector(
        num_classes=config.get('num_classes', 8)
    )
    model = model.to(args.device)
    
    print(f"✓ Model initialized")
    print("Note: Add your detection training logic here.")


def train_navigation_agent(args, config):
    """Train navigation agent."""
    print("🤖 Training Navigation Agent...")
    
    agent = NavigationAgent(
        state_dim=config.get('state_dim', 12),
        action_dim=config.get('action_dim', 6)
    )
    
    print(f"✓ Agent initialized")
    print("Note: Add your RL training logic here.")


def train_galaxy_classifier(args, config):
    """Train galaxy classifier."""
    print("🌌 Training Galaxy Classifier...")
    
    model = GalaxyClassifier(
        num_classes=config.get('num_classes', 5)
    )
    model = model.to(args.device)
    
    print(f"✓ Model initialized")
    print("Note: Add your training logic here.")


def train_exoplanet_detector(args, config):
    """Train exoplanet detector."""
    print("🪐 Training Exoplanet Detector...")
    
    model = ExoplanetDetector(
        sequence_length=config.get('sequence_length', 2000)
    )
    model = model.to(args.device)
    
    print(f"✓ Model initialized")
    print("Note: Add your training logic here.")


def main():
    """Main training function."""
    args = parse_args()
    config = load_config(args.config)
    
    print("=" * 60)
    print("🚀 AI Space Exploration - Model Training")
    print("=" * 60)
    
    # Select training function based on model
    train_functions = {
        'satellite_classifier': train_satellite_classifier,
        'object_detector': train_object_detector,
        'navigation_agent': train_navigation_agent,
        'galaxy_classifier': train_galaxy_classifier,
        'exoplanet_detector': train_exoplanet_detector,
    }
    
    train_fn = train_functions.get(args.model)
    if train_fn:
        train_fn(args, config)
    else:
        print(f"Error: Unknown model '{args.model}'")
        sys.exit(1)


if __name__ == "__main__":
    main()
